优化名片、企业、商品分类、商品等权限

+++ AiCard/AiCard/Controllers/ProductsManageController.cs
@@ -24,10 +24,10 @@ namespace AiCard.Controllers
         public void Sidebar(string name = "商品管理")
         {
             ViewBag.Sidebar = name;
-
         }
 
         // GET: ProductsManage
+        [Authorize(Roles = SysRole.ProductManageRead + "," + SysRole.EProductManageRead)]
         public ActionResult Index(string filter, int page = 1)
         {
             Sidebar();
@@ -52,7 +52,6 @@ namespace AiCard.Controllers
                         Kind = p.Kind,
                         Images=p.Images
                     };
-            //var m = db.Products.Include(ps => ps.Kind);
             if (!string.IsNullOrWhiteSpace(filter))
             {
                 m = m.Where(s => s.Name.Contains(filter));
@@ -63,7 +62,6 @@ namespace AiCard.Controllers
                 m = m.Where(s => s.EnterpriseID == AccontData.EnterpriseID);
             }
             var paged = m.OrderByDescending(s => s.ID).ToPagedList(page);
-           
             return View(paged);
         }
 
@@ -83,6 +81,7 @@ namespace AiCard.Controllers
         }
 
         // GET: ProductsManage/Create
+        [Authorize(Roles = SysRole.ProductManageCreate + "," + SysRole.EProductManageCreate)]
         public ActionResult Create()
         {
             ViewBag.KindID = new SelectList(db.ProductKinds, "ID", "Name");
@@ -95,8 +94,16 @@ namespace AiCard.Controllers
         // 详细信息，请参阅 http://go.microsoft.com/fwlink/?LinkId=317598。
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = SysRole.ProductManageCreate + "," + SysRole.EProductManageCreate)]
         public ActionResult Create(ProductViewModels product)
         {
+            var tempuser = db.Users.FirstOrDefault(s => s.Id == AccontData.UserID);
+            //防止企业用户串号修改
+            if (AccontData.UserType == Enums.UserType.Enterprise
+                && tempuser.EnterpriseID != AccontData.EnterpriseID)
+            {
+                return this.ToError("错误", "没有权限修改当前角色", Url.Action("Index"));
+            }
             if (ModelState.IsValid)
             {
                 var model = new Product
@@ -117,19 +124,25 @@ namespace AiCard.Controllers
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-
             ViewBag.KindID = new SelectList(db.ProductKinds, "ID", "Name", product.KindID);
             return View(product);
         }
 
         // GET: ProductsManage/Edit/5
+        [Authorize(Roles = SysRole.ProductManageEdit + "," + SysRole.EProductManageEdit)]
         public ActionResult Edit(int? id)
         {
+            var temp = db.Products.FirstOrDefault(s => s.ID == id);
+            //防止企业用户串号修改
+            if (AccontData.UserType == Enums.UserType.Enterprise
+                && temp.EnterpriseID != AccontData.EnterpriseID)
+            {
+                return this.ToError("错误", "没有权限修改当前角色", Url.Action("Index"));
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            //Product product = db.Products.Find(id);
             var model = (from p in db.Products
                          from ps in db.ProductKinds
                          where p.KindID == ps.ID && p.ID == id.Value
@@ -162,15 +175,24 @@ namespace AiCard.Controllers
         // 详细信息，请参阅 http://go.microsoft.com/fwlink/?LinkId=317598。
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [ValidateInput(false)]
+        [Authorize(Roles = SysRole.ProductManageEdit + "," + SysRole.EProductManageEdit)]
         public ActionResult Edit(ProductViewModels product)
         {
+            var temp = db.Products.FirstOrDefault(s => s.ID == product.ID);
+            //防止企业用户串号修改
+            if (AccontData.UserType == Enums.UserType.Enterprise
+                && temp.EnterpriseID != AccontData.EnterpriseID)
+            {
+                return this.ToError("错误", "没有权限修改当前角色", Url.Action("Index"));
+            }
             if (ModelState.IsValid)
             {
                 var t = db.Products.FirstOrDefault(s => s.ID == product.ID);
                 t.ID = product.ID;
                 t.Name = product.Name;
                 t.Count = product.Count;
-                t.EnterpriseID = product.EnterpriseID;
+                //t.EnterpriseID = AccontData.EnterpriseID;
                 t.Info = product.Info;
                 t.KindID = product.KindID;
                 t.Price = product.Price;
@@ -190,25 +212,66 @@ namespace AiCard.Controllers
         }
 
         // GET: ProductsManage/Delete/5
+        [Authorize(Roles = SysRole.ProductMangeDelete + "," + SysRole.EProductMangeDelete)]
         public ActionResult Delete(int? id)
         {
+            var temp = db.Products.FirstOrDefault(s => s.ID == id);
+            //防止企业用户串号修改
+            if (AccontData.UserType == Enums.UserType.Enterprise
+                && temp.EnterpriseID != AccontData.EnterpriseID)
+            {
+                return this.ToError("错误", "没有权限修改当前角色", Url.Action("Index"));
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Product product = db.Products.Find(id);
-            if (product == null)
+
+
+            var model = (from p in db.Products
+                         from ps in db.ProductKinds
+                         from e in db.Enterprises
+                         where p.KindID == ps.ID&&p.EnterpriseID==e.ID && p.ID == id.Value
+                         select new ProductIndexViewModels
+            {
+                ID = p.ID,
+                Name = p.Name,
+                Count = p.Count,
+                EnterpriseID = p.EnterpriseID,
+                EnterpriseName=e.Name,
+                Info = p.Info,
+                KindID = p.KindID,
+                KindName=ps.Name,
+                Price = p.Price,
+                Release = p.Release,
+                Sort = p.Sort,
+                TotalSales = p.TotalSales,
+                Type = p.Type
+            }).FirstOrDefault();
+            if (model == null)
             {
                 return HttpNotFound();
             }
-            return View(product);
+            return View(model);
+
+
+            //Product product = db.Products.Find(id);
+            return View(model);
         }
 
         // POST: ProductsManage/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = SysRole.ProductMangeDelete + "," + SysRole.EProductMangeDelete)]
         public ActionResult DeleteConfirmed(int id)
         {
+            var temp = db.Products.FirstOrDefault(s => s.ID == id);
+            //防止企业用户串号修改
+            if (AccontData.UserType == Enums.UserType.Enterprise
+                && temp.EnterpriseID != AccontData.EnterpriseID)
+            {
+                return this.ToError("错误", "没有权限修改当前角色", Url.Action("Index"));
+            }
             Product product = db.Products.Find(id);
             db.Products.Remove(product);
             db.SaveChanges();
